Make xcb-icccm dependency be optional and off-by-default
Changes to the xcb-icccm API/ABI are under discussion, so for now default
to using local property handling code modeled after the current API, with
a --with-xcb-icccm to enable use of the API for testing/development.

Once the API/ABI is stable & released, this set of changes should be
removed and the xcb-icccm API just used directly.

Change-Id: I8c04696be522684bbcd551aa90bb3e5939c73954
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Reviewed-by: James Cloos <cloos@jhcloos.com>